Add persistent user
- Add persistent login, when user closes app and reopen, the user does
not need to re-log
